Modificação do componente probabilidade para não aceitar letras nem números fora do intervalo de 0 a 100
Modificação do componente Impacto para apresentar uma opção padrão (--Selecione--)

Exclusão do pacote "Teste tabela"

Modificação do tipo de "Grau de Severidade" para Double

+++ src/view/RiscosPriorizarRiscosJPanel.java
@@ -400,7 +400,7 @@ public class RiscosPriorizarRiscosJPanel extends javax.swing.JPanel {
                     emissorJTextField.setText(RiscoSelecionado.getEmissor());
                     probabilidadeJTextField.setText(Integer.toString(RiscoSelecionado.getProbabilidade()));
                     impactoJTextField.setText((RiscoSelecionado.getImpacto()));
-                    grauDeSeveridadeJTextField.setText(Integer.toString(RiscoSelecionado.getGrauSeveridade()));
+                    grauDeSeveridadeJTextField.setText(Double.toString(RiscoSelecionado.getGrauSeveridade()));
 
                 }
